Updated logger format, now uses time since start of run.
